backend/requirements.txt
flask
flask-cors
joblib
torch
transformers
nltk
scikit-learn
requests
python-dotenv